+++ src/DemandFunction.cpp
@@ -2,7 +2,7 @@
 
 DemandFunction::DemandFunction() {
 	demandVector.clear();
-	nExp = 1.3;
+	nExp = 1.3f;
 }
 
 DemandFunction::~DemandFunction() {
@@ -25,7 +25,7 @@ float DemandFunction::addDemandsTogether(const float x, const float y) const {
 
 	float demand = 0;
 
-	for (int i = 0; i < demandVector.size(); ++i) {
+	for (unsigned long i = 0; i < demandVector.size(); ++i) {
 		
 		demand += getZforFunction(demandVector[i], x, y);
 
@@ -38,9 +38,9 @@ float DemandFunction::addDemandsTogether(const float x, const float y) const {
 
 float DemandFunction::getZforFunction(const DemandFunctionStruct & d, const float x, const float y) const {
 
-	float pxy = pow(fabs(x-d.x),2)+pow(fabs(y-d.y),2);
-	float nxy = pow(fabs(x-d.x),nExp)+pow(fabs(y-d.y),nExp);
-	if (nxy == 0) nxy = 0.0001;
+	float pxy = static_cast<float>(pow(fabs(x-d.x),2.f)+pow(fabs(y-d.y),2.f));
+	float nxy = static_cast<float>(pow(fabs(x-d.x),nExp)+pow(fabs(y-d.y),nExp));
+	if (nxy == 0) nxy = 0.0001f;
 
 	return getZforStruct(d, pxy, nxy);
